feat(scms): Phase 0 foundations - enums, decimal quantities, UoM conversion, document numbering, real user attribution
Task 1: xUnit test project at tests/api-scm.Tests running against real PostgreSQL, with
characterization tests pinning 24 pre-existing defects.
Task 2: status enums with a [DbValue] attribute preserving the exact legacy strings (zero schema
change), plus IStatusTransitionGuard state machines. Closes a hole where a transfer could go
Pending -> Completed and be marked delivered without debiting the source.
Task 3: all stock quantities int -> decimal(18,3). 0.75 kg is now representable.
Task 4: unit-of-measure conversion (UomType/ConversionFactor/BaseUom, Item.StockUomId). Closes the
thousand-fold error where a recipe in grams was subtracted from a kilogram balance.
Task 5: gapless document numbering (PO-2026-0001), IPostingTransaction so every stock posting is
atomic, unique index on Inventories(ItemId, LocationId), xmin concurrency token.
Task 6: ICurrentUserService reading auth-service claims; AuditLog/InventoryMovementLog UserId
int -> text so a real identity fits; removed every hardcoded UserId = 1.

111 tests passing.

+++ Api/Contracts/ProductionContracts.cs
@@ -4,7 +4,10 @@ public class CreateProductionBatchRequest
 {
     public int RecipeId { get; set; }
     public int ProductId { get; set; }
-    public int BatchMultiplier { get; set; }
+
+    /// <summary>How many standard recipe batches to cook. Decimal so a half batch is allowed.</summary>
+    public decimal BatchMultiplier { get; set; }
+
     public DateTime ScheduleDate { get; set; }
     public string AssignedCook { get; set; } = string.Empty;
 }
@@ -16,9 +19,9 @@ public class ProductionBatchResponse
     public string RecipeName { get; set; } = string.Empty;
     public int ProductId { get; set; }
     public string ProductName { get; set; } = string.Empty;
-    public int BatchMultiplier { get; set; }
-    public int EstimatedQuantity { get; set; }
-    public int ActualQuantity { get; set; }
+    public decimal BatchMultiplier { get; set; }
+    public decimal EstimatedQuantity { get; set; }
+    public decimal ActualQuantity { get; set; }
     public DateTime ProductionDate { get; set; }
     public string Stage { get; set; } = string.Empty;
     public string Status { get; set; } = string.Empty;
@@ -32,7 +35,7 @@ public class ProductionBatchResponse
 public class UpdateStageRequest
 {
     public string Stage { get; set; } = string.Empty;
-    public int? ActualQuantity { get; set; }
+    public decimal? ActualQuantity { get; set; }
 }
 
 public class UpdateQaNotesRequest
@@ -59,6 +62,6 @@ public class LowStockAlertResponse
 {
     public int ItemId { get; set; }
     public string ItemName { get; set; } = string.Empty;
-    public int CurrentStock { get; set; }
-    public int MinStockLevel { get; set; }
+    public decimal CurrentStock { get; set; }
+    public decimal MinStockLevel { get; set; }
 }
