+++ Api/Contracts/Requests/CreateStockTransferRequest.cs
@@ -5,6 +5,6 @@ public class CreateStockTransferRequest
     public int ProductId { get; set; }
     public int SourceLocationId { get; set; }
     public int DestLocationId { get; set; }
-    public int TransferQuantity { get; set; }
+    public decimal TransferQuantity { get; set; }
     public DateTime? TransferDate { get; set; }
 }
