+++ Domains/Entities/BatchConsumption.cs
@@ -5,7 +5,7 @@ public class BatchConsumption
     public int BatchConsumptionId { get; set; }
     public int BatchId { get; set; }
     public int ItemId { get; set; }
-    public int RequiredQuantity { get; set; }
+    public decimal RequiredQuantity { get; set; }
     public int SourceSupplierId { get; set; }
 
     public ProductionBatch? Batch { get; set; }
